Injector/Controller workers now work.
sm_consume and sm_inject commands now have a "--pure" option, which tells it
to use the pure gearmand client/worker interface and issue requests via the
injector/controller workers.

This all works, except now QueueRunner needs to be adjusted to properly handle
when to send requeued jobs through the injector worker.

May have to add a flags field to the DB to indicate what we want to do with
the job. Another (grosser?) option is to look for the 'inject_jobs' queue from
the gearman status command.