add 'flag' and 'failcount' field, use 'flag'
if injecting a job via the injector, add a flag to ensure job is re-injected
through the controller if needed. Use the shim by default.

Note that this sort of forces to you to *either* the shim or the
injector/controller. Not sure if it's necessary to allow a combination though.